% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{all_are}
\alias{all_are}
\title{all_are}
\usage{
all_are(inplist, typechar)
}
\arguments{
\item{inplist}{list of List to be checked}

\item{typechar}{char to check for}
}
\value{
boolean, for whether the elements of \code{inplist} are all \code{typechar}
}
\description{
Checks if all elements of a list or List are of a (single) particular type \code{typechar}
}
\examples{
x <- list(1,2)
all_are(x,'numeric')
all_are(x,'char')

y <- list(1,2,'c')
all_are(y,'numeric')
all_are(y,'char')
}
