% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePercentNotNull.R
\name{calcPercentNotNull}
\alias{calcPercentNotNull}
\title{Calculate percent not null}
\usage{
calcPercentNotNull(x)
}
\arguments{
\item{x}{Any data frame, matrix, vector.}
}
\value{
This function takes data (a vector, data frame, or matrix) and calculates the percentage of observations that are not \code{NA} or \code{NULL} values. This is used to calculate the percentage of loci with valid SNP data.
}
\description{
This function calculates the percentage of columns (loci) that are not \code{NA} or \code{NULL} values i.e., that have valid SNP data.
}
