% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corral.R
\name{compsvd}
\alias{compsvd}
\title{Compute Singular Value Decomposition (SVD)}
\usage{
compsvd(preproc_mat, method = c("irl", "svd")[1], ncomp = 10, ...)
}
\arguments{
\item{preproc_mat}{matrix, pre-processed input; can be sparse or full [pre-processing can be performed using corral_preproc or pca_preproc from this package]}

\item{method}{character, the algorithm to be used for svd. Default is irl. Currently supports 'irl' for irbla::irlba or 'svd' for stats::svd}

\item{ncomp}{numeric, number of components (if using irlba); Default is 10}

\item{...}{other parameters for the svd and irlba functions can also be utilized. See those documentation for options.}
}
\value{
SVD result. List of matrices, dvu, d is  diagonal matrix with the eigenvalues
}
\description{
Compute Singular Value Decomposition (SVD)
}
\examples{
compsvd(my_matrix)
compsvd(my_matrix, method = 'svd')
compsvd(my_matrix, method = 'irl', ncomp = 20)
}
