% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertBasePairstoCodes.R
\name{convertBasePairstoCodes}
\alias{convertBasePairstoCodes}
\title{Convert base pairs to single character IUPAC codes}
\usage{
convertBasePairstoCodes(initdata)
}
\arguments{
\item{initdata}{A data frame containing a column \code{Coral_ID} with unique value per row, as well as at least one column containing base pair data with at least one value matching valid allele data in the \code{IUPAC} table.}
}
\value{
This function returns a data frame identical to the input data frame except the paired alleles (e.g., G:) are replace by single allele codes (e.g., G). Note that only "allowable data" are included in the resulting data frame, and any columns not adhering to the data formatting requirements are omitted by \code{checkforAllowableData()} and listed in a warning issued by \code{handleError_ProhibitedData()}.
}
\description{
This function replaces the base pairs with \code{IUPAC} codes and converts the data to wide format so that there is a column per locus. If any invalid characters are present in a column, then the entire column is deleted, a warning is issued, and the column names are reported.
}
