% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corral.R
\name{corral}
\alias{corral}
\title{Correspondence analysis for a single matrix}
\usage{
corral(inp, ...)
}
\arguments{
\item{inp}{matrix (any type), or \code{SingleCellExperiment}. If using \code{SingleCellExperiment}, then include the \code{whichmat} argument to specify which slot to use (defaults to \code{counts}).}

\item{...}{}
}
\value{
For matrix input, returns list with the correspondence analysis matrix decomposition result (u,v,d are the raw svd output; SCu and SCv are the standard coordinates; PCu and PCv are the principal coordinates)
For SingleCellExperiment input, returns the SCE with embeddings in the reducedDim portion under 'corral'
}
\description{
This is a wrapper for \code{\link{corral_mat}} and \code{\link{corral_sce}}. See those functions for the list of possible parameters.
}
