% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corral.R
\name{corral_mat}
\alias{corral_mat}
\title{Correspondence analysis on a single matrix}
\usage{
corral_mat(inp_mat, method = c("irl", "svd")[1], ncomp = 10, ...)
}
\arguments{
\item{inp_mat}{matrix, raw or lognormed counts (no negative values)}

\item{method}{character, the algorithm to be used for svd. Default is irl. Currently supports 'irl' for irbla::irlba or 'svd' for stats::svd}

\item{ncomp}{numeric, number of components (if using irlba); Default is 10}
}
\value{
list with the correspondence analysis matrix decomposition result (u,v,d are the raw svd output; SCu and SCv are the standard coordinates; PCu and PCv are the principal coordinates)
}
\description{
Correspondence analysis on a single matrix
}
\examples{
result <- corral_mat(my_matrix)
result <- corral_mat(my_matrix, method = 'svd')
result <- corral_mat(my_matrix, method = 'irl', ncomp = 30)
}
