% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corral.R
\name{corral_sce}
\alias{corral_sce}
\title{Correspondence analysis on a single matrix (SingleCellExperiment)}
\usage{
corral_sce(
  inp_sce,
  method = c("irl", "svd")[1],
  ncomp = 10,
  whichmat = "counts",
  fullout = FALSE,
  ...
)
}
\arguments{
\item{inp_sce}{SingleCellExperiment; raw or lognormed counts (no negative values)}

\item{method}{character; the algorithm to be used for svd. Default is irl. Currently supports 'irl' for irbla::irlba or 'svd' for stats::svd}

\item{ncomp}{numeric; number of components (if using irlba); Default is 10}

\item{whichmat}{character; defaults to \code{counts}, can also use \code{logcounts} or \code{normcounts} if stored in the \code{sce} object}

\item{fullout}{boolean; whether the function will return the full \code{corral} output as a list, or a SingleCellExperiment; defaults to SingleCellExperiment (\code{FALSE}). To get back the \code{\link{corral_mat}}-style output, set this to \code{TRUE}.}
}
\value{
(default) SCE with the embeddings in the \code{reducedDim} slot \code{corral}. Also can return the same output as \code{\link{corral_mat}}.
}
\description{
Correspondence analysis on a single matrix (SingleCellExperiment)
}
\examples{
result <- corral_sce(my_sce)
result <- corral_sce(my_sce, method = 'svd')
result <- corral_sce(my_sce, method = 'irl', ncomp = 30, whichmat = 'logcounts')

# example on how to add UMAP based on corral above, with 'scater' package
library(scater)
result <- runUMAP(result, dimred = 'corral', name = 'corral_UMAP')
result <- runTSNE(result, dimred = 'corral', name = 'corral_TSNE')

}
