% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corralm.R
\name{corralm_sce}
\alias{corralm_sce}
\title{Multi-table correspondence analysis}
\usage{
corralm_sce(sce, splitby, method = c("irl", "svd")[1], ncomp = 10, ...)
}
\arguments{
\item{sce}{SingleCellExperiment; containing the data to be integrated. Default is to use the counts, and to include all of the data in the integration. These can be changed by passing additional arguments. See sce2matlist function documentation for list of available parameters.}

\item{splitby}{character; name of the attribute from colData that should be used to separate the SCE}

\item{method}{character; the algorithm to be used for svd. Default is irl. Currently supports 'irl' for irbla::irlba or 'svd' for stats::svd}

\item{ncomp}{numeric; number of components (if using irlba); Default is 10}

\item{...}{for additional parameters, see the sce2matlist function}
}
\value{
list with the correspondence analysis matrix decomposition result
}
\description{
Multi-table correspondence analysis
}
\examples{
result <- corralm_sce(my_sce, 'platform')

#' # example on how to add UMAP/tsne based on corralm above, with 'scater' package
library(scater)
result <- runUMAP(result, dimred = 'corralm', name = 'corralm_UMAP')
result <- runTSNE(result, dimred = 'corralm', name = 'corralm_TSNE')

}
