% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{earthmover_dist}
\alias{earthmover_dist}
\title{Earthmover distance
i.e., wasserstein distance with L1 (p_param = 1); can also use other penalties > 1
(Not technically earthmover distance if using other p_param values)}
\usage{
earthmover_dist(batch1, batch2, whichdim = 1, numbins = 100, p_param = 1)
}
\arguments{
\item{batch1}{matrix; probably a dimension from an embedding (must match in dim with batch2)}

\item{batch2}{matrix; probably a dimension from an embedding (must match in dim with batch1)}

\item{whichdim}{int; which dimension (i.e., column) from the embeddings is used. defaults on first}

\item{numbins}{int; number of bins for the probability discretization (defaults to 100)}

\item{p_param}{int; penalty parameter for wasserstein distance. Defaults to 1, which corresonds to earthmover.}
}
\value{
double; the distance
}
\description{
Earthmover distance
i.e., wasserstein distance with L1 (p_param = 1); can also use other penalties > 1
(Not technically earthmover distance if using other p_param values)
}
