% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_methods.R
\name{eval_perf}
\alias{eval_perf}
\title{Evaluate performance}
\usage{
eval_perf(routine)
}
\arguments{
\item{routine}{code to be evaluated}
}
\value{
list with 'time' and 'memory change' as elements
}
\description{
Evaluate performance
}
\examples{
perf_report <- eval_perf(my_corral <- corral(my_matrix))
}
