% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fix_na}
\alias{fix_na}
\title{Fix NA}
\usage{
fix_na(x)
}
\arguments{
\item{x}{matrix of values for which na values should be changed to 0}
}
\value{
fixed matrix
}
\description{
Fix NA
}
\examples{
fix_na(x)
}
