% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaled_variance.R
\name{get_var_for_attr}
\alias{get_var_for_attr}
\title{Get variance for observations of an attribute}
\usage{
get_var_for_attr(attr_name, attr_vec, embedding, pcnum)
}
\arguments{
\item{attr_name}{character; the name of the attribute to compute variance for}

\item{attr_vec}{vector of characters; the vector of values for that attribute corresponding to the rows in embedding}

\item{embedding}{matrix; the embeddings, where columns correspond to PCs and rows correspond to cells (entries)}

\item{pcnum}{int; the PC to compute the variance for}
}
\value{
double; the scaled variance for that subset of observations with attr_name
}
\description{
Get variance for observations of an attribute
}
