% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_w}
\alias{get_w}
\title{Get weights: computes row weights and column weights}
\usage{
get_w(inp_mat)
}
\arguments{
\item{inp_mat}{matrix for which weights should be calculated (sparse or full)}
}
\value{
list of 2 elements: 'row.w' and 'col.w' contain the row and column weights respectively
}
\description{
Get weights: computes row weights and column weights
}
\examples{
get_w(my_mat)
lapply(my_matlist, get_w)
}
