% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_weights}
\alias{get_weights}
\title{Get weights: computes row weights and column weights}
\usage{
get_weights(inp_mat)
}
\arguments{
\item{inp_mat}{matrix for which weights should be calculated (sparse or full)}
}
\value{
list of 2 elements: 'row.w' and 'col.w' contain the row and column weights respectively
}
\description{
Get weights: computes row weights and column weights
}
\examples{
mat <- matrix(sample(seq(0,20,1),100,replace = TRUE),nrow = 10)
ws <- get_weights(mat)
}
