% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorChecks.R
\name{handleError_ProhibitedData}
\alias{handleError_ProhibitedData}
\title{Notify user that non-conforming fields were found and omitted.}
\usage{
handleError_ProhibitedData(dataset, acceptableData)
}
\arguments{
\item{dataset}{A data frame to be tested with a \code{Coral_ID} column and SNP data, along with other fields such as site name etc.}

\item{acceptableData}{A data frame with acceptable data that is allowable; here, these are \code{IUPAC} SNP data.}
}
\value{
This function returns a warning message notifying users of the presence and identity of columns in their data that do not adhere to formatting requirements. This is used in conjunction with \code{checkforAllowableData()}, which uses this function and also omits the offending columns.
}
\description{
This function notifies users that non-conforming fields/columns such as site name or, more importantly, allele pairs that are not included in the \code{IUPAC} reference file, were omitted from the data set. This function omits the offending columns and reports a message that they were removed along with their names.
}
