% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list2mat}
\alias{list2mat}
\title{List to Matrix}
\usage{
list2mat(matlist, direction = c("c", "r")[1])
}
\arguments{
\item{matlist}{list of matrices to concatenate}

\item{direction}{character, r or c, to indicate whether should be row-wise (i.e., rbind to match on columns) or column-wise (i.e., cbind to match on rows). Defaults to columnwise (matching on rows) to match convention of SingleCellExperiments}
}
\value{
matrix
}
\description{
List to Matrix
}
\examples{
list2mat(my_matrix_list)
list2mat(my_matrix_list, 'r')
}
