% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{na2zero}
\alias{na2zero}
\title{Set na to 0}
\usage{
na2zero(x)
}
\arguments{
\item{x}{matrix of values for which na values should be changed to 0}
}
\value{
matrix, where na values are set to 0
}
\description{
Set na to 0
}
\examples{
x <- matrix(sample(0:10, 5000, replace = TRUE), ncol = 25)
x[sample(1:5000, 10)] <- NA

na2zero(x)
}
