% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omitInvariantLoci.R
\name{omitInvariantLoci}
\alias{omitInvariantLoci}
\title{Identify and omit invariant loci for subsequent kinship calculations.}
\usage{
omitInvariantLoci(dataset)
}
\arguments{
\item{dataset}{A data frame with \code{1:nrow(dataset)} individual colonies as rows and loci with SNP data in \code{2:ncol(dataset)} columns.}
}
\value{
This function returns the supplied data frame with invariant loci (i.e., columns) removed. This is done for subsequent kinship calculations that must exclude invariant loci (see \code{runKinship()} and \code{kinshipCalcsNoInvar()}).
}
\description{
This function omits invariant loci for subsequent population- and individual-level kinship calculations.
}
