% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pairwise_rv}
\alias{pairwise_rv}
\title{Pairwise rv coefficient}
\usage{
pairwise_rv(matlist)
}
\arguments{
\item{matlist}{list of matrices (or matrix-like; see \code{rv} function) for which to compute pairwise RV coefficients}
}
\value{
matrix of the pairwise coefficients
}
\description{
Pairwise rv coefficient
}
\examples{
a <- matrix(sample(1:10,100, TRUE), nrow = 10)
b <- matrix(sample(1:10,50, TRUE), nrow = 5)
c <- matrix(sample(1:10,20, TRUE), nrow = 2)

matlist <- list(a,b,c)
pairwise_rv(matlist)
pairwise_rv(lapply(matlist, t))
}
