% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{pca_preproc}
\alias{pca_preproc}
\title{Principal components analysis (PCA) preprocessing}
\usage{
pca_preproc(inp_mat, type = "corr")
}
\arguments{
\item{inp_mat}{matrix of logcounts; can be sparse or full}

\item{type}{character, default, runs correlation-based (using 'corr' argument); can also run covariance-based (using 'cov' argument)}
}
\value{
preprocessed sparse matrix upon which compsvd can be run to complete PCA routine
}
\description{
Principal components analysis (PCA) preprocessing
}
\examples{
mat_pca <- pca_preproc(my_matrix, type = 'cov')
pca_output <- compsvd(mat_pca)
}
