% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_embedding.R
\name{plot_embedding}
\alias{plot_embedding}
\title{Plot selected PCs from an embedding}
\usage{
plot_embedding(
  embedding,
  xpc = 1,
  ypc = xpc + 1,
  plot_title = paste0("PC", xpc, " by PC", ypc),
  color_vec,
  color_title,
  ellipse_vec = NULL,
  saveplot = TRUE,
  plotfn = paste(plot_title, xpc, sep = "_"),
  showplot = TRUE,
  returngg = FALSE
)
}
\arguments{
\item{embedding}{matrix or other tabular format where columns correspond to PCs and rows correspond to cells (entries)}

\item{xpc}{int; which PC to put on the x-axis (defaults to 1)}

\item{ypc}{int; which PC to put on the y-axis (defaults to the one after xpc)}

\item{plot_title}{char; title of plot (defaults to titling based on xpc and ypc)}

\item{color_vec}{vector; length should correspond to the number of rows in embedding, and each element of the vector classifies that cell (entry) in the embedding to that particular class, which will be colored the same. (e.g., this could be indicating which batch each cell is from)}

\item{color_title}{char; what attribute the colors represent}

\item{ellipse_vec}{OPTIONAL vector; length should correspond to the number of rows in embedding, and each element of the vector classifies that cell (entry) in the embedding to that particular class, and elements of the same class will be circled in an ellipse. (e.g., this could be indicating the cell type or cell line; works best for attributes intended to be compact)}

\item{saveplot}{boolean; whether or not to save the plot, defaults TRUE}

\item{plotfn}{char; what the filename is to be called. (defaults to making a name based on plot title and xpc)}

\item{showplot}{boolean; whether or not to show the plot, defaults TRUE}

\item{returngg}{boolean; whether or not to return a ggplot2 object, defaults FALSE}
}
\value{
default none; options to display plot (showplot), save plot (saveplot), and/or return ggplot2 object (returngg)
}
\description{
Plot selected PCs from an embedding
}
\examples{

embed_mat <- corralm(list_of_mats)
plot_embedding(embedding = embed_mat, xpc = 1, plot_title = 'corralm plot',color_vec = cell_type_vec, color_title = 'cell type')

}
