% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnGenetIdentity.R
\name{returnGenetIdentity}
\alias{returnGenetIdentity}
\title{Determine the genet identity of each coral colony based on pairwise comparisons of SNPs}
\usage{
returnGenetIdentity(obs)
}
\arguments{
\item{obs}{a data frame created by \code{groupByGenets()} that has eight columns: \code{coral1}, \code{coral2}, \code{CoralPair}, \code{pctMatch}, \code{pctNotNull}, \code{PartOfGenet}, \code{AdequateData}, and \code{obs}. \code{coral1} identifies the first coral in the pair, \code{coral2} identifies the second coral in the pair, and \code{CoralPair} identifies the pair, taking the format of the first coral identity concatenated together with the second coral identity, separated by a period. \code{pctMatch} and \code{pctNotNull} summarize the alleles for the comparison being made, \code{PartOfGenet} means \code{pctMatch} exceeds a threshold for identifying individuals as a match or clone, \code{AdequateData} characterizes the frequency of \code{NULL} or \code{NA} values, and \code{obs} is simply a sequential observation number that is used to group genets.}
}
\value{
This function returns a data frame with each colony assigned to a numeric genet number. Multiple colonies can belong to the same genet, but all corals are compared to themselves; hence, each colony is classified as a clone of itself by default, provided it had adequate SNP data. and they therefore occupy their own genet and are unrelated to any other colonies. Colonies previously determined to have insufficient data for this classification are not included in this analysis and are instead added to the genet assignment data frame at a later step with \code{genet = XXXX_NA}, where \code{XXXX} is the 4-letter species code.
}
\description{
This function assigns coral colonies to genets based on all possible comparisons (i.e., comparing colonies to themselves and other colonies) genetic similarity of SNPs at each of N loci.
}
