% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rv}
\alias{rv}
\title{rv coefficient}
\usage{
rv(mat1, mat2)
}
\arguments{
\item{mat1}{matrix (or matrix-like, e.g., df); either columns or rows should be matched with \code{mat2}}

\item{mat2}{matrix (or matrix-like, e.g., df); either columns or rows should be matched with \code{mat1}}
}
\value{
numeric; RV coefficient between the matched matrices
}
\description{
rv coefficient
}
\examples{
a <- matrix(sample(1:10,100, TRUE), nrow = 10)
b <- matrix(sample(1:10,50, TRUE), nrow = 5)

rv(a, b) # matched by columns
rv(t(a), t(b)) # matched by rows
}
