% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaled_variance.R
\name{scaled_var_comparison}
\alias{scaled_var_comparison}
\title{Compute scaled variances by attribute to compare multiple embeddings on a single PC}
\usage{
scaled_var_comparison(embeds_list, attr_vec, pcnum = 1)
}
\arguments{
\item{embeds_list}{list of matrices; each matrix is an embedding, where columns correspond to PCs and rows correspond to cells (entries). Ideally the embeddings are named appropriately, as names for the output will be assigned based on names of the list elements.}

\item{attr_vec}{attr_vec vector of characters; the vector of values for that attribute corresponding to the rows in embedding}

\item{pcnum}{int; the PC to compute the scaled variance for}
}
\value{
data.frame; columns correspond to different embeddings from the list and the rows correspond to each unique attribute type in attr_vec for that embedding
}
\description{
Compute scaled variances by attribute to compare multiple embeddings on a single PC
}
\examples{
embed_list <- list('corralm' = reducedDim(sce,'corralm'), 'corral' = reducedDim(sce,'corral'))
scaled_var_comparison(embed_list, colData(sce)$Method, 2)
}
