% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaled_variance.R
\name{scaled_var_df_by_attr}
\alias{scaled_var_df_by_attr}
\title{Compute scaled variance by attribute for a single embedding, multiple PCs}
\usage{
scaled_var_df_by_attr(embedding, attr_vec, pcnums = c(1:3))
}
\arguments{
\item{embedding}{matrix; the embeddings, where columns correspond to PCs and rows correspond to cells (entries)}

\item{attr_vec}{vector of characters; the vector of values for that attribute corresponding to the rows in embedding}

\item{pcnums}{vector; the PCs to compute the variances for}
}
\value{
data.frame; columns correspond to PCs and the rows correspond to each unique attribute type in attr_vec for this embedding, on that PC
}
\description{
Compute scaled variance by attribute for a single embedding, multiple PCs
}
