% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaled_variance.R
\name{scaled_var_vec_by_attr}
\alias{scaled_var_vec_by_attr}
\title{Compute scaled variance by attribute for a single embedding}
\usage{
scaled_var_vec_by_attr(embedding, attr_vec, pcnum = 1)
}
\arguments{
\item{embedding}{matrix; the embeddings, where columns correspond to PCs and rows correspond to cells (entries)}

\item{attr_vec}{vector of characters; the vector of values for that attribute corresponding to the rows in embedding}

\item{pcnum}{int; the PC to compute the scaled variance for}
}
\value{
vector; scaled variances for each unique attribute type in attr_vec for this embedding
}
\description{
Compute scaled variance by attribute for a single embedding
}
