% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_embeddings2scelist}
\alias{add_embeddings2scelist}
\title{Add embeddings to SCE}
\usage{
add_embeddings2scelist(scelist, embeddings, slotname = "corral")
}
\arguments{
\item{scelist}{list of SingleCellExperiments; to which the corresponding embeddings should be added}

\item{embeddings}{matrix; the embeddings outputted from a dimension reduction, e.g. \code{\link{corralm}}. Rows in this table correspond to columns in the SCEs in \code{scelist} (if all the SCEs were column-bound), and row indices should correspond to cells.}

\item{slotname}{character; name of the slot for the reduced dim embedding; defaults to \code{corral}}
}
\value{
list of SingleCellExperiments with respective embeddings stored in them
}
\description{
Add embeddings to SCE
}
