% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corral.R
\name{corral_preproc}
\alias{corral_preproc}
\title{Correspondence analysis preprocessing (not intended to be called directly)}
\usage{
corral_preproc(inp_mat, rtype = c("standardized", "indexed")[1])
}
\arguments{
\item{inp_mat}{matrix, counts or logcounts; can be sparse or full}

\item{rtype}{character indicating what type of residual should be computed; options are "indexed" and "standardized"; defaults to 'standardized'}
}
\value{
sparse matrix, preprocessed, upon which compsvd can be run to complete CA routine
}
\description{
Correspondence analysis preprocessing (not intended to be called directly)
}
\examples{
mat_corral <- corral_preproc(my_matrix)
corral_output <- compsvd(mat_corral)
}
