% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corral.R
\name{corral_sce}
\alias{corral_sce}
\title{Correspondence analysis on a single matrix (SingleCellExperiment)}
\usage{
corral_sce(
  inp_sce,
  method = c("irl", "svd")[1],
  ncomp = 10,
  whichmat = counts,
  ...
)
}
\arguments{
\item{inp_sce}{SingleCellExperiment; raw or lognormed counts (no negative values)}

\item{method}{character; the algorithm to be used for svd. Default is irl. Currently supports 'irl' for irbla::irlba or 'svd' for stats::svd}

\item{ncomp}{numeric; number of components (if using irlba); Default is 10}

\item{whichmat}{accessor function; defaults to counts, can also use logcounts or normcounts if available}
}
\value{
SCE with the SVD output in the reducedDim portion under 'corral'
}
\description{
Correspondence analysis on a single matrix (SingleCellExperiment)
}
\examples{
result <- corral_sce(my_sce)
result <- corral_sce(my_sce, method = 'svd')
result <- corral_sce(my_sce, method = 'irl', ncomp = 30, whichmat = logcounts)

# example on how to add UMAP based on corral above, with 'scater' package
library(scater)
result <- runUMAP(result, dimred = 'corral', name = 'corral_UMAP')
result <- runTSNE(result, dimred = 'corral', name = 'corral_TSNE')

}
