% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corralm.R
\name{corralm}
\alias{corralm}
\title{Multi-table correspondence analysis}
\usage{
corralm(inp, ...)
}
\arguments{
\item{inp}{list of matrices (any type), \code{SingleCellExperiment}, or list of \code{SingleCellExperiment}s. If using \code{SingleCellExperiment}, then include the \code{whichmat} argument to specify which slot to use (defaults to \code{counts}).}

\item{...}{}
}
\value{
For a list of matrices input, returns list with the correspondence analysis matrix decomposition result
For SingleCellExperiment input, returns the SCE with embeddings in the reducedDim slot 'corral'
For a list of SingleCellExperiments, returns a list of the SCEs with the embeddings in the respective reducedDim slot 'corral'
}
\description{
This is a wrapper for \code{\link{corralm_matlist}} and \code{\link{corral_sce}}. See those functions for the list of possible parameters.
}
