% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_pct_var_exp_svd}
\alias{get_pct_var_exp_svd}
\title{Compute percent of variance explained}
\usage{
get_pct_var_exp_svd(thissvd, preproc_mat = thissvd$d)
}
\arguments{
\item{thissvd}{list outputted from svd function (svd or irlba)}

\item{preproc_mat}{matrix of pre-processed values (optional) - important to include if the svd is only partial as this is used to compute the sum of eigenvalues}
}
\value{

}
\description{
Compute percent of variance explained
}
\examples{
get_pct_var_exp_svd(my_svd) # this works if my_svd is a full svd
get_pct_var_exp_svd(my_svd, preproc_mat = my_mat) # ... otherwise use this 
}
