% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_embedding.R
\name{plot_embedding_sce}
\alias{plot_embedding_sce}
\title{Plot selected PCs from an embedding saved in a SingleCellExperiment object}
\usage{
plot_embedding_sce(
  sce,
  which_embedding,
  color_attr,
  color_title = color_attr,
  ellipse_attr = NULL,
  ...
)
}
\arguments{
\item{sce}{\code{\link{SingleCellExperiment}} object; contains the embedding within the \code{reducedDim} slot}

\item{which_embedding}{character; for the embedding to plot}

\item{color_attr}{character; name of the attribute within \code{colData} to use for assigning colors (in lieu of \code{color_vec} in the \code{\link{plot_embedding}} function)}

\item{color_title}{character; title to use for colors legend, defaults to the same as \code{color_attr}}

\item{ellipse_attr}{OPTIONAL character; name of the attribute within \code{colData} to use for drawing ellipse(s) (in lieu of \code{ellipse_vec} in the \code{\link{plot_embedding}} function)}

\item{...}{additional optional arguments - see \code{\link{plot_embedding}} function for details on other potential arguments: \code{xpc}, \code{ypc}, \code{plot_title}, \code{color_title} (if title is different from \code{color_attr}), \code{saveplot}, \code{plotfn}, \code{showplot}, \code{returngg}}
}
\value{
default none; options to display plot (\code{showplot}), save plot (\code{saveplot}), and/or return \code{\link{ggplot2}} object (\code{returngg})
}
\description{
Plot selected PCs from an embedding saved in a SingleCellExperiment object
}
\examples{

sce <- corralm(sce, 'Method')

# to plot and show only
plot_embedding_sce(sce = sce, which_embedding = 'corralm', xpc = 1, plot_title = 'corralm: PC1 by PC2',color_attr = "Method", ellipse_attr = 'cell_type')

# to return ggplot2 object and display, but not save
corralm_ggplot <- plot_embedding_sce(sce = sce, which_embedding = 'corralm', xpc = 1, plot_title = 'corralm: PC1 by PC2',color_attr = 'Method', ellipse_attr = 'cell_type', returngg = TRUE, saveplot = FALSE)

# or, to plot UMAP, using 'scater' package:
library(scater)
result <- runUMAP(sce, dimred = 'corralm', name = 'corralm_UMAP')
plot_embedding_sce(sce = result, which_embedding = 'corralm_UMAP', xpc = 1, plot_title = 'corralm UMAP: PC1 by PC2',color_attr = "Method", ellipse_attr = 'cell_type')

}
